requirements.txt
pandas>=1.5.0
numpy>=1.21.0
ccxt>=4.0.0
matplotlib>=3.5.0
pytest>=7.0.0
python-dotenv>=1.0.0
requests>=2.28.0

# Optional integrations
twilio>=8.0.0  # For SMS integration
